requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dash==3.2.0
dash-bootstrap-components==2.0.4
Flask==3.1.2
gunicorn==23.0.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
narwhals==2.10.1
nest-asyncio==1.6.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
plotly==6.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
retrying==1.4.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
